pyproject.toml
[tool.poetry]
name = "langcards"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.4"
psycopg2 = "^2.9.7"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
flake8 = "^6.1.0"
Pillow = "^10.0.0"
pytest-django = "^4.5.2"
pytest-cov = "^4.1.0"
factory-boy = "^3.3.0"
requests = "^2.31.0"
ipython = "^8.15.0"
google-api-python-client = "^2.98.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
blurhash-python = "^1.2.1"
tqdm = "^4.66.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
